+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Remove `Context`. Use `Input` instead.
+- Remove the implementation of `PartialEq` and `Eq` for `Snapshot` and `Instant` to prevent unintended string comparison.
 
 ## [0.7.0] - 2025-02-16